// frontend/src/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Any: { input: any; output: any; }
  /**
   * Define a Relay Cursor type:
   * https://relay.dev/graphql/connections.htm#sec-Cursor
   */
  Cursor: { input: any; output: any; }
  Duration: { input: any; output: any; }
  Map: { input: any; output: any; }
  Time: { input: any; output: any; }
  UUID: { input: any; output: any; }
  Upload: { input: any; output: any; }
};

export type Action = Node & {
  __typename?: 'Action';
  actionFiles?: Maybe<Array<InvocationFiles>>;
  bazelInvocation: BazelInvocation;
  commandLine?: Maybe<Array<Scalars['String']['output']>>;
  configuration: Configuration;
  endTime?: Maybe<Scalars['Time']['output']>;
  exitCode?: Maybe<Scalars['Int']['output']>;
  failureCode?: Maybe<Scalars['String']['output']>;
  failureMessage?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  label: Scalars['String']['output'];
  startTime?: Maybe<Scalars['Time']['output']>;
  stderrHash?: Maybe<Scalars['String']['output']>;
  stderrHashFunction?: Maybe<Scalars['String']['output']>;
  stderrSizeBytes?: Maybe<Scalars['Int']['output']>;
  stdoutHash?: Maybe<Scalars['String']['output']>;
  stdoutHashFunction?: Maybe<Scalars['String']['output']>;
  stdoutSizeBytes?: Maybe<Scalars['Int']['output']>;
  success?: Maybe<Scalars['Boolean']['output']>;
  type?: Maybe<Scalars['String']['output']>;
};
// ...
export type BazelInvocation = Node & {
  __typename?: 'BazelInvocation';
  actions?: Maybe<Array<Action>>;
  authenticatedUser?: Maybe<AuthenticatedUser>;
  bazelVersion?: Maybe<Scalars['String']['output']>;
  bepCompleted: Scalars['Boolean']['output'];
  build?: Maybe<Build>;
  /** JSON representation of the canonical command line options. */
  canonicalCommandLine?: Maybe<Scalars['Map']['output']>;
  configurations?: Maybe<Array<Configuration>>;
  connectionMetadata?: Maybe<ConnectionMetadata>;
  endedAt?: Maybe<Scalars['Time']['output']>;
  environmentVariables?: Maybe<Scalars['Map']['output']>;
  exitCodeCode?: Maybe<Scalars['Int']['output']>;
  exitCodeName?: Maybe<Scalars['String']['output']>;
  hostname?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  instanceName: InstanceName;
  invocationFiles?: Maybe<Array<InvocationFiles>>;
  invocationID: Scalars['UUID']['output'];
  invocationTargets: InvocationTargetConnection;
  metrics?: Maybe<Metrics>;
  numFetches?: Maybe<Scalars['Int']['output']>;
  /** JSON representation of the parsed command line options */
  optionsParsed?: Maybe<Scalars['Map']['output']>;
  /** JSON representation of the original command line options. */
  originalCommandLine?: Maybe<Scalars['Map']['output']>;
  profile?: Maybe<Profile>;
  sourceControl?: Maybe<Array<SourceControl>>;
  startedAt?: Maybe<Scalars['Time']['output']>;
  tags: InvocationTagConnection;
  username?: Maybe<Scalars['String']['output']>;
};
// ...
export type Configuration = Node & {
  __typename?: 'Configuration';
  actions?: Maybe<Array<Action>>;
  bazelInvocation: BazelInvocation;
  configurationID: Scalars['String']['output'];
  cpu?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  invocationTargets?: Maybe<Array<InvocationTarget>>;
  isTool?: Maybe<Scalars['Boolean']['output']>;
  makeVariables?: Maybe<Scalars['Map']['output']>;
  mnemonic?: Maybe<Scalars['String']['output']>;
  platformName?: Maybe<Scalars['String']['output']>;
};
// ...
export type InvocationFiles = Node & {
  __typename?: 'InvocationFiles';
  action?: Maybe<Action>;
  bazelInvocation?: Maybe<BazelInvocation>;
  content?: Maybe<Scalars['String']['output']>;
  digest?: Maybe<Scalars['String']['output']>;
  digestFunction?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  invocationTarget?: Maybe<InvocationTarget>;
  name: Scalars['String']['output'];
  sizeBytes?: Maybe<Scalars['Int']['output']>;
  uri?: Maybe<Scalars['String']['output']>;
};
// ...
export type InvocationTarget = Node & {
  __typename?: 'InvocationTarget';
  abortReason: InvocationTargetAbortReason;
  bazelInvocation: BazelInvocation;
  configuration?: Maybe<Configuration>;
  durationInMs?: Maybe<Scalars['Int']['output']>;
  endTimeInMs?: Maybe<Scalars['Int']['output']>;
  failureMessage?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  startTimeInMs?: Maybe<Scalars['Int']['output']>;
  success: Scalars['Boolean']['output'];
  tags?: Maybe<Array<Scalars['String']['output']>>;
  target: Target;
  targetFiles?: Maybe<Array<InvocationFiles>>;
  testSummary?: Maybe<Array<TestSummary>>;
};

/** InvocationTargetAbortReason is enum for the field abort_reason */
export enum InvocationTargetAbortReason {
  AnalysisFailure = 'ANALYSIS_FAILURE',
  Incomplete = 'INCOMPLETE',
  Internal = 'INTERNAL',
  LoadingFailure = 'LOADING_FAILURE',
  None = 'NONE',
  NoAnalyze = 'NO_ANALYZE',
  NoBuild = 'NO_BUILD',
  OutOfMemory = 'OUT_OF_MEMORY',
  RemoteEnvironmentFailure = 'REMOTE_ENVIRONMENT_FAILURE',
  Skipped = 'SKIPPED',
  TimeOut = 'TIME_OUT',
  Unknown = 'UNKNOWN',
  UserInterrupted = 'USER_INTERRUPTED'
}
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
};
// ...
export type BazelInvocationActionsFragment = { __typename?: 'Action', id: string, label: string, type?: string | null, success?: boolean | null, exitCode?: number | null, commandLine?: Array<string> | null, startTime?: any | null, endTime?: any | null, failureCode?: string | null, failureMessage?: string | null, stdoutHash?: string | null, stdoutSizeBytes?: number | null, stdoutHashFunction?: string | null, stderrHash?: string | null, stderrSizeBytes?: number | null, stderrHashFunction?: string | null, actionFiles?: Array<{ __typename?: 'InvocationFiles', name: string, uri?: string | null }> | null, configuration: { __typename?: 'Configuration', id: string, configurationID: string, mnemonic?: string | null, platformName?: string | null, cpu?: string | null, makeVariables?: any | null } } & { ' $fragmentName'?: 'BazelInvocationActionsFragment' };